begun implementing the functions needed for shipping example

+++ PathFinding/Shipping/Types.fs
@@ -2,6 +2,6 @@
 
 open PathFinding
 
-type Harbor = { id : int; maxShips : int; fee : int; pos : Position; }
+type Harbor = { id : int; maxShips : int; fee : float; pos : Position; }
 type Container = { id : int; dest : int; cargoType : int; }
-type Route = { from : int; dest : int; cost : int; }
+type Route = { from : int; dest : int; cost : float; }
